rename observer folder
