+++ src/backoffice/editors/shared/node/editor-node.element.ts
@@ -7,7 +7,7 @@ import { UmbNodeStore } from '../../../../core/stores/node.store';
 import { NodeEntity } from '../../../../core/mocks/data/node.data';
 import type { UmbNotificationService } from '../../../../core/services/notification';
 import { UmbNotificationDefaultData } from '../../../../core/services/notification/layouts/default';
-import { UmbObserverMixin } from '../../../../core/observer';
+import { UmbObserverMixin } from '../../../../core/observable-api';
 import { UmbNodeContext } from './node.context';
 import { UmbContextConsumerMixin, UmbContextProviderMixin } from '@umbraco-cms/context-api';
