+++ src/backoffice/editors/shared/node/views/edit/editor-view-node-edit.element.ts
@@ -7,7 +7,7 @@ import { UmbNodeContext } from '../../node.context';
 import { UmbContextConsumerMixin } from '@umbraco-cms/context-api';
 
 import '../../../../../components/node-property/node-property.element';
-import { UmbObserverMixin } from '../../../../../../core/observer';
+import { UmbObserverMixin } from '../../../../../../core/observable-api';
 
 @customElement('umb-editor-view-node-edit')
 export class UmbEditorViewNodeEditElement extends UmbContextConsumerMixin(UmbObserverMixin(LitElement)) {
