+++ src/backoffice/trees/shared/tree-navigator.element.ts
@@ -3,7 +3,7 @@ import { repeat } from 'lit/directives/repeat.js';
 import { UUITextStyles } from '@umbraco-ui/uui-css/lib';
 import { customElement, state } from 'lit/decorators.js';
 import { Entity } from '../../../core/mocks/data/entities';
-import { UmbObserverMixin } from '../../../core/observer';
+import { UmbObserverMixin } from '../../../core/observable-api';
 import { UmbTreeDataContextBase } from '../tree-data.context';
 import { UmbContextConsumerMixin } from '@umbraco-cms/context-api';
